backend/pyproject.toml
[project]
name = "CogniCube_backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "XuChenXu", email = "91937041+ChenXu233@users.noreply.github.com"},
]
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
    "sqlalchemy>=2.0.38",
    "databases>=0.9.0", 
    "google-auth>=2.38.0", 
    "gspread>=6.1.4", 
    "fastapi-mail>=1.4.2", 
    "google-auth-oauthlib>=1.2.1", 
    "google-auth-httplib2>=0.2.0"
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
start = "pdm run app.py"